\name{pawacc-package}
\alias{pawacc-package}
\alias{pawacc}
\docType{package}
\title{
Physical Activity with Accelerometers
}
\description{
This package provides processing and summary functions.
}
\details{
\tabular{ll}{
Package: \tab pawacc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-04\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Marco Geraci

Maintainer: Marco Geraci <m.geraci@ucl.ac.uk>
}

\keyword{ package }
\keyword{ physical activity }
\keyword{ accelerometer }
