\name{pawacc-package}
\alias{pawacc-package}
\alias{pawacc}
\docType{package}
\title{
Physical Activity with Accelerometers
}
\description{
This package provides processing and summary functions.
}
\details{
\tabular{ll}{
Package: \tab pawacc\cr
Type: \tab Package\cr
Version: \tab 1.2.1\cr
Date: \tab 2014-12-10\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}

\references{
Geraci M, Rich C, Sera F, Cortina-Borja M, Griffiths LJ, and Dezateux C (2012). Technical report on accelerometry data processing in the Millennium Cohort Study. London, UK: University College London. Available at \url{http://discovery.ucl.ac.uk/1361699}
}

\author{
Marco Geraci

Maintainer: Marco Geraci <m.geraci@ucl.ac.uk>
}

\keyword{ package }
\keyword{ physical activity }
\keyword{ accelerometer }
