\name{print.acclist}
\alias{print.accfile}
\alias{print.acclist}

\title{
Print an \code{accfile} or \code{acclist} Object
}
\description{
Print an object generated by \code{\link{gt1m.accfile}}, \code{\link{gt3x.accfile}} or \code{\link{readAccDir}}.
}

\usage{
\method{print}{acclist}(x, ...)
\method{print}{accfile}(x, ...)
}

\arguments{
  \item{x}{
an \code{accfile} or an \code{acclist} object.
}
  \item{\dots}{
not used.
}
}

\author{
Marco Geraci
}

\keyword{ print }

