\name{print.acclist}
\alias{print.acclist}

\title{
Print an \code{acclist} Object
}
\description{
Print an object generated by \code{\link{gt1m.acclist}}.
}

\usage{
\method{print}{acclist}(x, ...)
}

\arguments{
  \item{x}{
an \code{acclist} object.
}
  \item{\dots}{
not used.
}
}

\author{
Marco Geraci
}

\keyword{ print }

